Refactor ONNX chat server into dedicated package
