+++ apps/onnx_chat/main.py
@@ -1,13 +1,14 @@
-import os
+"""Minimal OpenAI-compatible chat server using FastAPI."""
+
 from typing import List, Dict, Any
+import os
 
 from fastapi import FastAPI, HTTPException, Request
-from fastapi.responses import JSONResponse
 from pydantic import BaseModel, ValidationError
 
 try:
-    import onnxruntime as ort
-except Exception:  # pragma: no cover - optional dependency
+    import onnxruntime as ort  # optional
+except Exception:  # pragma: no cover
     ort = None
 
 app = FastAPI()
@@ -22,15 +23,16 @@ MODEL_PATH = os.environ.get("ONNX_MODEL_PATH")
 
 
 def load_session():
+    """Load the ONNX model on first use if available."""
     global session
     if session is None and ort and MODEL_PATH and os.path.exists(MODEL_PATH):
         session = ort.InferenceSession(MODEL_PATH)
 
 
 def classify(text: str) -> str:
+    """Return a simple classification using the model or a rule based fallback."""
     load_session()
     if session is None:
-        # Fallback simple rule when ONNX model is unavailable
         return "positive" if "good" in text.lower() else "negative"
     inputs = {session.get_inputs()[0].name: [[ord(c) for c in text]]}
     outputs = session.run(None, inputs)[0]
@@ -41,13 +43,9 @@ def classify(text: str) -> str:
 async def chat(request: Request):
     try:
         data = await request.json()
-    except Exception:
-        raise HTTPException(status_code=400, detail="Invalid JSON")
-
-    try:
         req = ChatRequest(**data)
-    except ValidationError as e:
-        raise HTTPException(status_code=400, detail=e.errors())
+    except Exception as e:  # JSON error or validation
+        raise HTTPException(status_code=400, detail=str(e))
 
     text = ""
     if req.messages:
@@ -58,6 +56,7 @@ async def chat(request: Request):
                     text += part.get("text", "")
         else:
             text = str(content)
+
     result = classify(text)
     return {
         "id": "cmpl-001",
@@ -69,11 +68,12 @@ async def chat(request: Request):
     }
 
 
-def main():
+def main():  # pragma: no cover
     import uvicorn
 
-    uvicorn.run(app, host="0.0.0.0", port=int(os.environ.get("PORT", "8000")))
+    log_level = os.getenv("ONNX_CHAT_LOG_LEVEL", "info")
+    uvicorn.run(app, host="0.0.0.0", port=int(os.environ.get("PORT", "8000")), log_level=log_level)
 
 
-if __name__ == "__main__":
+if __name__ == "__main__":  # pragma: no cover
     main()
