+++ apps/onnx_chat/requirements.txt
@@ -1,0 +1,4 @@
+fastapi
+uvicorn
+pydantic
+onnxruntime
